rename component package to defaultcomponent
