+++ Library_System/src/view/advertising/AdImageButton.java
@@ -12,7 +12,7 @@ import javax.swing.SwingConstants;
 import javax.swing.border.LineBorder;
 
 import view.advertising.details.AdDetailsView;
-import view.component.DefaultFrame;
+import view.defaultcomponent.DefaultFrame;
 	/*
 	 	[AdImageButton]
 	 	버튼에 들어갈 이미지와 텍스트 사이즈를 조정해주는 클래스.
