+++ Library_System/src/view/defaultcomponent/BookListWithSelectedBook.java
@@ -1,4 +1,4 @@
-package view.component;
+package view.defaultcomponent;
 
 import java.util.ArrayList;
