+++ Library_System/src/view/defaultcomponent/DefaultBookSearchBar.java
@@ -1,4 +1,4 @@
-package view.component;
+package view.defaultcomponent;
 
 import java.awt.Dimension;
 import java.awt.Font;
