+++ Library_System/src/view/defaultcomponent/IFrame.java
@@ -1,4 +1,4 @@
-package view;
+package view.defaultcomponent;
 
 public interface IFrame {
