+++ Library_System/src/view/main/MainFrame.java
@@ -1,18 +1,30 @@
 package view.main;
 
-import view.component.DefaultFrame;
+import view.defaultcomponent.DefaultFrame;
 
 
 public class MainFrame extends DefaultFrame{
 	
 	public MainFrame() {
+
+		super();
+		setFrame();
 		
 		setLayout(null);
 		
+		setComp();
+		setDesign();
+	}
+	
+	@Override
+	public void setComp() {
+
+	}
+
+	@Override
+	public void setDesign() {
 		add(new TopPanel());
 		add(new CenterPanel());
-		
-		setVisible(true);
 		
 	}
 	
@@ -25,5 +37,7 @@ public class MainFrame extends DefaultFrame{
 		});
 
 	}
+
+	
 
 }
