+++ Library_System/src/view/main/TopPanel.java
@@ -7,7 +7,7 @@ import java.awt.event.ActionListener;
 import javax.swing.JButton;
 import javax.swing.JFrame;
 
-import view.component.DefaultPanel;
+import view.defaultcomponent.DefaultPanel;
 import view.login.LoginFrame;
 
 public class TopPanel extends DefaultPanel{
