+++ Library_System/src/view/manager/dialog/ManagerDialog.java
@@ -30,11 +30,11 @@ import javax.swing.border.EmptyBorder;
 
 import model.dto.AllBookInfo;
 import model.dto.Bookinfo;
-import view.IFrame;
-import view.component.BookListWithSelectedBook;
-import view.component.DefaultButton;
-import view.component.DefaultPanel;
-import view.component.RoundTextField;
+import view.defaultcomponent.BookListWithSelectedBook;
+import view.defaultcomponent.DefaultButton;
+import view.defaultcomponent.DefaultPanel;
+import view.defaultcomponent.IFrame;
+import view.defaultcomponent.RoundTextField;
 
 
 public class ManagerDialog extends JDialog implements IFrame {
